+++ extensions/hue-palette/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Hue Changelog
 
+## [Update API Endpoint] - {PR_MERGE_DATE}
+
+- Update the hue color image API endpoint to `hue-palette.yencheng.dev`.
+
 ## [Maintenance] - 2025-09-18
 
 - Add support for Windows platform.